requirements.txt
pandas>=2.0,<3.0
numpy>=1.24,<2.0
yfinance>=0.2.40,<0.3
requests
python-dotenv
APScheduler>=3.10,<4.0
